Rename `protocol.py` to `tcp.py`
